payments/serializers.py
from rest_framework import serializers

from payments.models import Order, OrderItem

class OrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Order
        fields = '__all__'