requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
Django==4.2.18
django-cors-headers==4.5.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
mysqlclient==2.2.6
pillow==11.0.0
psycopg-binary==3.1.19
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.0
python-dotenv==1.0.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.2
stripe==11.3.0
typing_extensions==4.12.2
urllib3==2.2.3
